Add a few more tests, move tests
